Add validations in coworking model

+++ src/routes/deleteCoworking.js
@@ -4,7 +4,11 @@ module.exports = (app) => {
     app.delete('/api/coworkings/:id', (req, res) => {
         Coworking.findByPk(req.params.id)
             .then(coworking => {
-                Coworking.destroy({
+                if (coworking === null) {
+                    const message = `Le coworking demandé n'existe pas.`
+                    return res.status(404).json({ message })
+                }
+                return Coworking.destroy({
                     where: {
                         id: req.params.id
                     }
@@ -13,6 +17,10 @@ module.exports = (app) => {
                         const message = `Le coworking ${coworking.name} a bien été supprimé.`
                         res.json({ message, data: coworking });
                     })
+            })
+            .catch(error => {
+                const message = `Impossible de supprimer le coworking.`
+                res.status(500).json({ message, data: error })
             })
     });
 }
